Attempt to extend short paths that go nowhere
If a short path has one end that is not shared with another path, then
we can probably extend that end a little to make the path longer so as
to allow a longer name.  The name will fall off the end of the path, but
is unlikely to intersect something else in a problematic way.

If neither end is connected to some other path, we can extend in both
directions.

Signed-off-by: NeilBrown <neil@brown.name>